refactor : NumberGenerator의 역할을 더 자연스럽게 인터페이스로 추상화하여라
- 기존의 구조인, 추상 클래스가 random number 메서드를 가진것이 자연스럽지 않았다.
- 만약 상속받은 메서드가 없다면, 굳이 추상 클래스로 만들 필요가 없다고 생각해 인터페이스로 추상화하였다.
- MockNumberGenerator, RandomUniqueNumberGenerator로 더 자연스럽게 분리되었다.
- LottoMachine도 어떻게 번호가 추출되는지 알 필요 없이, 주입된 NumberGenerator로 번호만 뽑아 생성하면 되도록 하여 좀 더 자연스러워졌다.
