+++ src/main/java/generator/NumberGenerator.java
@@ -1,0 +1,7 @@
+package generator;
+
+import java.util.List;
+
+public interface NumberGenerator {
+    List<Integer> generateNumbers(int count);
+}
