--- src/main/java/generator/RandomNumberGenerator.java
@@ -1,28 +1,0 @@
-package generator;
-
-import constants.ErrorType;
-import java.util.List;
-
-public abstract class RandomNumberGenerator {
-
-    final int startInclusive;
-    final int endInclusive;
-
-    public RandomNumberGenerator(final int startInclusive, final int endInclusive) {
-        validateStartLessThanEnd(startInclusive, endInclusive);
-        this.startInclusive = startInclusive;
-        this.endInclusive = endInclusive;
-    }
-
-    private void validateStartLessThanEnd(final int startInclusive, final int endInclusive) {
-        if (startInclusive > endInclusive) {
-            throw new IllegalArgumentException(ErrorType.NUMBER_GENERATOR_RANGE.getMessage());
-        }
-    }
-
-    public abstract List<Integer> pickNumbersInRange(final int count);
-
-    int generateRandomNumber() {
-        return (int) (Math.random() * endInclusive + startInclusive);
-    }
-}
