+++ src/main/java/generator/RandomUniqueNumberGenerator.java
@@ -4,18 +4,29 @@ import constants.ErrorType;
 import java.util.ArrayList;
 import java.util.List;
 
-public class RandomUniqueNumberGenerator extends RandomNumberGenerator {
+public class RandomUniqueNumberGenerator implements NumberGenerator {
+
+    final int startInclusive;
+    final int endInclusive;
 
     public RandomUniqueNumberGenerator(final int startInclusive, final int endInclusive) {
-        super(startInclusive, endInclusive);
+        validateStartLessThanEnd(startInclusive, endInclusive);
+        this.startInclusive = startInclusive;
+        this.endInclusive = endInclusive;
+    }
+
+    private void validateStartLessThanEnd(final int startInclusive, final int endInclusive) {
+        if (startInclusive > endInclusive) {
+            throw new IllegalArgumentException(ErrorType.NUMBER_GENERATOR_RANGE.getMessage());
+        }
     }
 
     @Override
-    public List<Integer> pickNumbersInRange(final int count) {
+    public List<Integer> generateNumbers(final int count) {
         validateCount(count);
         final List<Integer> randoms = new ArrayList<>();
         while (randoms.size() < count) {
-            int randomNumber = super.generateRandomNumber();
+            int randomNumber = generateRandomNumber();
             if (!randoms.contains(randomNumber)) {
                 randoms.add(randomNumber);
             }
@@ -25,10 +36,14 @@ public class RandomUniqueNumberGenerator extends RandomNumberGenerator {
     }
 
     private void validateCount(final int count) {
-        final int rangeCount = super.endInclusive - super.startInclusive + 1;
+        final int rangeCount = this.endInclusive - this.startInclusive + 1;
         if (rangeCount < count) {
             throw new IllegalArgumentException(ErrorType.RANDOM_NUMBER_GENERATOR_COUNT.getMessage());
         }
 
+    }
+
+    int generateRandomNumber() {
+        return (int) (Math.random() * endInclusive + startInclusive);
     }
 }
