+++ src/test/java/mock/MockNumberGenerator.java
@@ -1,18 +1,13 @@
 package mock;
 
-import generator.RandomNumberGenerator;
+import generator.NumberGenerator;
 import java.util.List;
 import java.util.stream.IntStream;
-import model.Lotto;
 
-public class MockNumberGenerator extends RandomNumberGenerator {
-
-    public MockNumberGenerator() {
-        super(Lotto.LOTTO_NUMBER_START_INCLUSIVE, Lotto.LOTTO_NUMBER_END_INCLUSIVE);
-    }
+public class MockNumberGenerator implements NumberGenerator {
 
     @Override
-    public List<Integer> pickNumbersInRange(final int count) {
+    public List<Integer> generateNumbers(final int count) {
         return IntStream.rangeClosed(1, count)
                 .boxed()
                 .toList();
